feat: nav bar 적용
