+++ src/components/@common/NavBar/NavBar.style.ts
@@ -1,4 +1,4 @@
-import { Link } from 'react-router-dom';
+import Link from 'next/link';
 import { styled } from 'styled-components';
 
 export const Wrapper = styled.nav`
